requirements.txt
africastalking==1.2.4
asgiref==3.4.1
catfish==1.4.10
certifi==2018.8.24
cffi==1.15.0
chardet==3.0.4
click==8.0.3
contextlib2==21.6.0
cryptography==36.0.1
cupshelpers==1.0
defusedxml==0.5.0
distro==1.3.0
distro-info==0.21
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
environs==9.3.5
evdev==1.1.2
Flask==2.0.2
gpg==1.12.0
gTTS==2.2.3
idna==2.6
img2pdf==0.3.3
importlib-metadata==4.8.2
inflect==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
lightdm-gtk-greeter-settings==1.2.2
lxml==4.3.2
MarkupSafe==2.0.1
marshmallow==3.14.1
mugshot==0.4.1
oauthlib==3.1.1
onboard==1.4.1
packaging==19.0
pdfarranger==1.6.2
pexpect==4.6.0
pikepdf==1.0.5+dfsg
Pillow==5.4.1
playsound==1.3.0
psycopg2==2.7.7
psycopg2-binary==2.9.2
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.30.4
PyJWT==2.3.0
pyparsing==2.2.0
pysmbc==1.0.15.6
PySocks==1.6.8
python-apt==1.8.4.1
python-debian==0.1.35
python-dotenv==0.19.2
python-xlib==0.23
python3-openid==3.2.0
pyttsx3==2.90
pytz==2021.3
pyudev==0.21.0
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.3.0
rpl==1.5.6
schema==0.7.5
six==1.12.0
sqlparse==0.4.2
torbrowser-launcher==0.3.3
typing_extensions==4.0.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
Werkzeug==2.0.2
whitenoise==5.3.0
zipp==3.6.0
